Changed Button Background Color
Changed button background color to Transparent. On this program's prior iteration, there was a lag between the change in the body's background color and the change in the button's background color, very noticable when changing from a dark color to a light one. By making the button transparent, there is no more lag.